9am push

+++ src/pages/Task.jsx
@@ -8,6 +8,7 @@ import { fetchAllTasks, deleteTask } from "../apis/apiHandlers";
 import CreateEditTask from "../forms/CreateEditTask";
 import ModalComponent from "../components/Modal";
 import TaskDetails from "../components/TaskDetails";
+import swal from "sweetalert";
 
 const Task = () => {
   const [show, setShow] = useState(false);
@@ -15,6 +16,7 @@ const Task = () => {
   const [error, setError] = useState(false);
 
   const [tasks, setTasks] = useState([]);
+  const [filteredTasks, setFilteredTasks] = useState([]);
   const [selectRow, setSelectRow] = useState(null);
   const [editRecord, setEditRecords] = useState(null);
 
@@ -24,12 +26,26 @@ const Task = () => {
       console.log("response = ", response);
 
       setTasks(response.data);
+      setFilteredTasks(response.data);
       setLoading(false);
     } catch (error) {
       console.log("error = ", error);
       toast.error("Network error");
       setLoading(false);
       setError(true);
+    }
+  };
+
+  const filter = (e) => {
+    if (e.target.value === "") {
+      setFilteredTasks(tasks);
+    } else {
+      let filtered = tasks.filter(
+        (item) =>
+          item.title.toLowerCase().includes(e.target.value.toLowerCase()) ||
+          item.description.toLowerCase().includes(e.target.value.toLowerCase())
+      );
+      setFilteredTasks(filtered);
     }
   };
 
@@ -48,16 +64,31 @@ const Task = () => {
     setSelectRow(record);
   };
 
-  const handleDeleteTask = async (id) => {
-    await deleteTask(id);
-    await getTasks();
-  };
 
   const handleEditTask = async (task) => {
     setEditRecords(task);
     handleShow();
   };
 
+  const handleDeleteTask = async (id) => {
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover the data!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then(async (willDelete) => {
+      if (willDelete) {
+        await deleteTask(id);
+        await getTasks();
+
+        swal("Customer Deleted Successfully!", {
+          icon: "success",
+        });
+      }
+    });
+  };
+console.log(filteredTasks);
   return (
     <div className="tasks">
       <div className="container-md">
@@ -71,24 +102,9 @@ const Task = () => {
               type="search"
               placeholder="Search Task"
               aria-label="Search"
+              onChange={filter}
             ></input>
-            <button
-              className="btn btn-outline-success  ms-2"
-              type="button"
-              onClick={() => {
-                toast.success("🦄 Wow so easy!", {
-                  position: "top-right",
-                  autoClose: 5000,
-                  hideProgressBar: false,
-                  closeOnClick: true,
-                  pauseOnHover: false,
-                  draggable: true,
-                  theme: "light",
-                });
-              }}
-            >
-              Search
-            </button>
+            {/* <button className="btn btn-outline-success  ms-2">Search</button> */}
           </form>
         </div>
         {loading ? (
@@ -105,16 +121,16 @@ const Task = () => {
           <div className="d-flex justify-content-center align-items-center">
             <h5>Network error</h5>
           </div>
-        ) : tasks?.length > 0 ? (
+        ) : filteredTasks?.length > 0 ? (
           <div className="table-responsive">
             <table className="table">
               <thead>
                 <tr>
-                  <th scope="col">Customer_ID</th>
+                  <th scope="col">Customer Phone No</th>
                   <th scope="col">Task title</th>
                   <th scope="col">Description</th>
                   <th scope="col">Address</th>
-                  {/* <th scope="col">Phone No</th> */}
+                  <th scope="col">Status</th>
                   <th scope="col" className="text-center" colSpan={2}>
                     Actions
                   </th>
@@ -125,7 +141,7 @@ const Task = () => {
                   handleDetailsModal={showDetailsModal}
                   onDeleteTask={handleDeleteTask}
                   onEditTask={handleEditTask}
-                  data={tasks}
+                  data={filteredTasks}
                 />
               </tbody>
             </table>
